refactor(zig-runtime): continue gameplay f64 to f32 narrowing (#273)
* refactor(zig): centralize creature lifecycle semantics

* refactor(zig-runtime): convert creature lifecycle state to f32

* refactor(zig-runtime): narrow bonus and spawn slot timers to f32

* refactor(runtime): keep projectile type ids typed in weapon fire paths

* refactor(runtime): remove projectile fallback defaults

* refactor(runtime): drop explicit invalid-id guards

* refactor(runtime): make weapon id conversion fail-fast

* refactor: hard cutover projectile meta to travel budget

* refactor: remove zig verifier legacy scaffolding

* refactor(zig-runtime): narrow creature damage and reward paths to f32

* refactor(zig-runtime): narrow creature formation math to f32

* refactor(zig-runtime): narrow creature death flow dt and world size to f32

* refactor(zig-runtime): narrow projectile damage interfaces to f32

* chore(gitignore): ignore local ast-grep config

* refactor(zig-runtime): remove redundant f32 narrowing in creature and particle paths

* refactor(zig-runtime): narrow particle runtime interfaces to f32

* refactor(zig-runtime): narrow secondary projectile interfaces to f32

* refactor(zig-runtime): narrow secondary detonation grid math to f32

* refactor(zig-runtime): narrow projectile spawn angles to f32

* refactor(zig-runtime): narrow creature spawn template stats to f32

* chore(deps): add numpy as a dev dependency

* refactor(zig-runtime): align projectile update f32 boundaries with native

* fix(runtime): avoid awarding xp to non-player owners

* refactor(runtime): key weapon and perk state by enums

* refactor(runtime): type most-used weapon lookup as WeaponId

* refactor(projectiles): align python projectile update with native f32 boundaries

* refactor(zig-runtime): advance f32 gameplay dt conversion

* refactor(replay): narrow quantize and timer helpers to f32

* refactor(zig-runtime): continue f32 narrowing in replay and weapons

* refactor(replay): narrow player movement dt paths to f32

* refactor(spawn): narrow quest timers and slot dt to f32

* refactor(bonuses): narrow gameplay bonus math to f32

* refactor(perks): remove remaining runtime f64 locals

* refactor(runtime): narrow distance helpers to f32

* refactor(projectiles): narrow non-trig scalars to f32

* refactor(runtime): narrow find-radius helper inputs to f32

* ci(zig): replace removed wasm smoke script

+++ crimson-zig/src/runtime/helpers.zig
@@ -24,8 +24,8 @@ pub fn directionFromAngle(angle: f32) state_mod.Vec2 {
 pub fn withinNativeFindRadius(
     origin: state_mod.Vec2,
     target: state_mod.Vec2,
-    radius: f64,
-    target_size: f64,
+    radius: f32,
+    target_size: f32,
 ) bool {
     const dx = target.x - origin.x;
     const dy = target.y - origin.y;
@@ -36,13 +36,13 @@ pub fn withinNativeFindRadius(
     return margin < 0.0;
 }
 
-pub fn distanceSq(a: state_mod.Vec2, b: state_mod.Vec2) f64 {
+pub fn distanceSq(a: state_mod.Vec2, b: state_mod.Vec2) f32 {
     const dx = a.x - b.x;
     const dy = a.y - b.y;
     return dx * dx + dy * dy;
 }
 
-pub fn distanceSqRoundedF32(a: state_mod.Vec2, b: state_mod.Vec2) f64 {
+pub fn distanceSqRoundedF32(a: state_mod.Vec2, b: state_mod.Vec2) f32 {
     return narrowF32(distanceSq(a, b));
 }
